Changes for TimeMonitor output.
I added the Evaluation Type to all evaluator names, so
that the timings are separated for Residual vs Jacobian
fill, for example.

I also made changes throughout Trilinos so that the
TimeMonitor output can be controlled for each of
the packages that use this function. The changes to
file doc/trilinos-cmake reflect this.