Rework of State Variable capabilty.
Added a Albany StateManager class that handles the
registration and allocation of storage for states.

During problem construction, state variables are
registered with a name and DataLayout.
    stateMgr.registerStateVariable("J2StressState",qp_tensor);

The allocation stage is performed once, and allocates
both new and old states for all worksets. Being the
third refacrtor of thuis capability this week, I hope
it is now done for a while!