Updated Albany for addition of stochastic parallelism support.
I updated Albany for the recent changes in Stokhos to support
parallelism over stochastic blocks.  This required some reengineering of
the solver factory to pass the right MPI comm's around correctly.
However stochastic parallelism doesn't quite work yet as there are still
a few MPI_COMM_WORLD's lying around that Andy has promised to fix.
Currently you specify the number of spatial processors through the
input file (which defaults to all of the processors supplied through
mpirun), but I hope to change this to a command line argument in the
future.

I also fixed all of the input files for the Teuchos Array-to-XML bug.