Add a test in for 2D J2 Plasticity, fix bug
I fixed a bug in the J2 model that was corrupting the calculation of Fp.
I added a test for the J2 stress model in the J2Plasticity2D directory.

Also, I get an overwhleming number of warning about comparing std::size_t
with and int, which happens all over the place.  I converted int to std::size_t
in a few places to get rid of a very small number of those warnings, but they
irritate me, so I'll probably change them in other places as well.

Lastly, I got some initialization warnings in a few constructors, so I fixed
those as well.