fix NonlinearElasticity 3D test
Recover from breaking that test with an earlier push.  Sorry about that.

Also, first shot at a volumetric averging within the deformation gradient.
Very quick and dirty for now, until I can figure out how to get the right info
from Intrepid to integrate properly.  But that's OK for now because it's as yet
untested and unused.