fix Neohookean and J2 bugs
There was a small error in the calculation of the shear modulus in both the
Neohookean and J2 models, easily caught when we started writing the stress
out to the output file.

The change causes a fraction of a percent difference in the average solution
response in the J2 model, so that value is updated (even though for some reason
that test was failing anyway, it should be passing now).