Added Charge Distribution output for QCAD problems.
The charge distribution is output, one value per element,
to the exodus file. THe name is currently hardwired to "state".

Stress output still works the same for NonlinearElasticity
problems. I added the logic to detect whether a state is
scalar,vector, or tensor, but have not added capability
to output multiple states.