Changed Problem specification to have "Solution Method" flag.
Previously, there were "Transient" and "Continuation" boolean
options in the input file. It wasn't clear what it meant for
them both to be true. So, these are no longer valid and
instead there is a "Solution Method" sttrrigthat can take
values of "Transient" "Continuation" or "Steady" (default).

Throughout the code, I made it the default to have time
dependent terms added to PDEs. This was previously tied
to the "transient" flag, but this blocked the code from
doing stability analysis with Shift-and-invert. So,
the ModelEvaluator supports transients even for
steady and continuation problems, which allows for
the eigensolver to probe the time dependent behavior.