Add capability to write out the residual (pre-Dicichlet BCs)
Using the STK interface to add a residual_field, this commit now allows for the
residual vector to be output before the application of the Dirichlet BCs.  This
is useful for the mechanics problems in calculating resultant forces and
similar things.

This adds some code to Albany_Application that I don't particularly like,
specifically the #ifdef and a cast.  When I figure out a better way to do
this, we can refactor the code.