Added new multi-point evaluation type.
This can be used to propagate multiple samples of a code simultaneously.
So far this is only used for an intrusive form of "nonintrusive
polynomial chaos" where all of the samples are propagated at once which
custom linear solvers to solve the resulting systems.  This takes
advantage of the vectorization of the residual/Jacobian evaluation, and
allows for basis recyling in the linear solvers.  An example/test was
added to test its functionality.