Added cmake options to allowing linking to LAME.
Example:

-D ENABLE_LAME:BOOL=ON \
-D LAME_INCLUDE_DIR:PATH=/Users/djlittl/Lame/include \
-D LAME_LIBRARY_DIR:PATH=/Users/djlittl/Lame/DEBUG \

I tried this out by hacking a call to the LAME
elastic material model in Stress_Def.hpp, which was
able to compile, link, and spit out correct results
for a simple call to getStress().  The code is
still there but commented out (marked DJL Hack).

The game plan is to essentially duplicate the
existing stress evaluator and, in the new copy,
hook things up to call LAME.

Then I'll go back and remove the junk
from Stress_Def.hpp.