Groundwork for hooking up LAME material models.
Created a LameStress evaluator that will call getStress in
the LAME material library.  It is currently not functional
but much of the plumbing is in place.

Modified CMakeLists.txt files to build new evaluator.  If
ENABLE_LAME is not set, a no-op version of the evaluator
will be built and will throw an exception if called.