Add bch, log_rotation, polar_left_logV, tests, and comments in LameStress_Def
The bch expansion is a method to approximate the log.
The log_rotation computes the log of a rotation via an explicit formula.
The polar_left_logV returns not only V and R, but also logV (the purpose of this
is to avoid recalling eig_spd).
Some more tests were added for these new methods.
I added to the comments in LameStress_Def.hpp, hopefully showing how to use the
Tensor class to get each of the desired quantities.