Progress on LameStress evaluator.
Hooked up new and improved functions from
the tensor class to fill the lame::matParams
structure.

The (currently inactive) test for LameStress
does not yet pass, so there's debugging to
be done.

In the process of hooking things up I was
getting compiler errors related to the
following tensor functions:

inverse(Tensor<ScalarT> const & A)
log_rotation(Tensor<ScalarT> const & R)
bch(Tensor<ScalarT> const & v, Tensor<ScalarT> const & r)

I believe the compiler errors were related
to the type of ScalarT the LameStress
evaluator is instantiated with.  I was
able to 'fix' the errors by explicitly
casting a few quantities to ScalarT in
the functions above.  I'm not sure if this
is the best fix, maybe the fix should actually
be done on the multiplication and division
operators for Tensor<ScalarT>.