Debugged the LameStress evaluator.  It now passes a modified version of the StaticElasticity3D test in which the Lame elastic material is called and the matrix-free solver is used.  This test is checked in under the name LameStaticElasticity3D. This test should only run if you compile with ENABLE_LAME.
Performance (the 'P' word) for the new test is very
bad compared to the standard StaticElasticity3D test,
even when the standard test is switched to use the matrix-free
solver.  For this reason the number of elements in the Lame
version of the test was reduced to 4x4x4.  There are several
obvious ways to improved performace of the LameStress
evaluator, so that's next on the list.  Hopefully these
changes will bring the Lame and non-Lame versions of
the test into the same ballpark performance-wise.

The log_rotation() tensor utility was modified to
include bounds checking on the argument to acos().  There
was numerical noise in the rotation matrix computed by
LameStress that caused the argument to be outside the
required -1 to 1 bounds, which was leading to a acos()
return value of NaN.  Values outside the acceptable range
are now mapped to -1 or 1 as needed.