Add nonlinear hardening, return mapping algorithm to J2 Plasticity
The J2Stress evaluator, which has linear hardening, was modified to accept two
additional parameters, a "Saturation Modulus" and a "Saturation Exponent" which
in effect extend to hardening of the model into the fully nonlinear regime.
The saturation type hardening model is presetned by Simo is a number of places,
but is still really just a toy hardening model as far as I am concerned.  The
return mapping algorithm was updated to reflect the nonlinear model, and most
importantly, everything still works because the defaults for these new
parameters are zero, which recovers linear hardening.

The two new hardning parameters have also been added in a "Sacadoized" fashion,
in fact a sensitivity study would be really interesting someday.

Add this to the list of features I need to verify.....