Fix to J2Stress, work on J2Damage, more cast issues with Tensor class
In the J2Stress model I missed a change necessary for the nonlinear hardening
to work properly.  Fixed that.

Also, I worked on converting the J2Damage model over to use the Tensor class.
Note that this means the GradientDamageProblem is only appropriate in 3D.
In doing so, I ran into some more of the ScalarT cast issues Dave Littlewood
was seeing.  For the time being, to get the code to compile, I simply casted
terms in a similar manner.  I'll follow up with Eric and Dave this week to
figure out the "right" thing to do.