Add Localization evaluator as a placeholder for localization elements
Localization elements are simply collapsed hex/wedge elements, consisting of
conincident surfaces in the reference configuration, that define the kinematics
through the relative displacement of the two surfaces.  They are used for
explicitly modeling cracks in solids and other localization phenomena.

Hopefully this will enable a collaboration with John Emery.