Doxygenized geometry and Zoltan partitioning utilities.
Fixed Doxygen directives so that there won't be duplicated
descriptions of member functions in the future.

The convention is that Doxygen documentation resides in the header files.

Added Doxygen configuration file Doxyfile that produces the documentation.
By executing doxygen in the utils directory both HTML and LaTeX documentation
will be produced.