Implementation of K field boundary condition
A K field BC is a Dirichlet BC that serves to apply a constant stress intensity
to a sharp crack geometry.  This implementation borrows heavily from the user
subroutine 'K_disp' in Sierra/SM.  A representative mesh and the input are
provided in the accompanying example.

But all is not well!  Parallel runs give bad results!  For that reason, the
example doesn't actually check the response, so this is really just a test
to see that it runs.  When I debug the parallel issue, I'll set the test up
(more) correctly.