Progress towards the refactor of STK usage.
There has been on ongoing design issue between the
mesh description and the problem setup. It is a
circular dependency (chicken-and-egg) issue, where each
depends on data from the other. In this commit, I've
broken the STK mesh generation into a 2 stage process.
I now need to propagate this through to the Problem
class.  Still needs more work.

Partition test: Added input.e to Partition test directory.
Specified in CMakeLists.txt for this data file to be copied
to test directory.