NOTE: Changes in this commit require a fresh pull of Trilinos.
Moved the "Workset Size" input file specification into the
Discretization sublist. All example problems with this line
were changed.

Added logic to adjust the Workset Size down to the problem
size on each processor, if it was larger. Also, added algorithm
to adjust the workset size down to minimize padding.
Since the workset size needs to be known before the mesh
is loaded, this requires a query of the mesh size from
stk::metaData before stk::bulkData is constructed.
Added a new method in stk::io to get this info.

Broke the discretization "create" method into two parts,
createMeshSpecs and createDiscretization. This is part
of an ongoing refactor.