Added two structures to communicate between the Albany::Problem and STK::Mesh. The MeshSpecsStruct contains the Mesh info needed to set up a Problem, and the StateInfoStruct contains the problem information needed to declare the Fields in a mesh.
Changes to all problems to use the minimal MeshSpecsStruct instead
of the  Discretization oject. The Discretization is now created
after the problem, so it can know about states.

Using StateInfoStruct to pass info through from the StateManager
to STK mesh.  Added getStateInfoStruct method to state manager to
parse the Field information.

Added NOX test for Finite Value in GradientDamage example, so
the code quits as soon as it hits inf/NaN.

Removed BoundaryFlux1DResponse since it was not general, and
did not fit with the new workflow (the problem is setup
before the mesh now).
- Removed  BoundaryFlux1D Response from input files.
- Re-derived the Dakota example problems to work without Boundary1D Flux Response.