Fix computation of dislocation density tensor, enable adaptive stepping
The dislocation density tensor needs to do some small, dense linear algebra
computations, so the Teuchos::SerialDenseMatrix/Solver are used.  I am
currently throwing an exception when the number of nodes and QPs are not the
same.  I'd like to use tets for this, so I'll get back to that eventually.

Also, the DefGrad evaluator, when the avgJ or volavgJ flags were set, would
test for a negative J.  Negative J causes NaNs in the stress calculation. So by
removing this check and adding in a FiniteValue test in the nonlinear solver
we can get adpative stepping on negative J.  Cool.