Major overhaul of handling of States.
All quantities that need their Old values to be saved from
one step to the next, and all quantities to be written to
Exodus, are deemed "states". These persist outside of a
single solve, and single workset, and are allocated memory
in the STK database.

There was an issue where both the Problem specification
and the STK Mesh construction needed information from the
other. THis has been solved as follows:
The setup phase between an Albany problem and STK_mesh now
goes like this:
(1) A small amount of mesh information is gathered and put
in  MeshSpecsStruct.
(2) The Albany Problem is set up completely, using some info
in MeshSpecsStruct. States are declared.
(3) A StateInfoStruct is passed back to STK that includes all
necessary information on states.
(4) The Fields are declared on the mesh (including states)
and the entire mesh is now loaded.

States are no longer averaged over elements, but
written out at every quadrature point. This needs to
be addressed again somehow:  averaging in Albany, averaging
in exodus using the algebra post-processing tool,
projection to nodes.

A Problem class just registers the stat