Progress toward properly initializing LAME material model state variables.
StateManager::initializeStateVariables() now accepts an init_type of, for example, "initial value 1.2",
where the 1.2 can be any scalar.

Created utility functions in Albany_Utils.* for converting doubles to/from strings corresponding to
this init_type, and also a function for checking the validity of string arguments for this init_type.

Created a utility function in LameUtils.* that calls a given LAME material model's initialize() function,
which returns initial values for state_old.

Plumbing is in place, testing is next.