add built-in potential calculations in the PoissonDirichlet evaluator and test the code for different temperatures, doping types, doping values, and statistics. Changes are made in QCAD_PoissonDirichelet.hpp and QCAD_PoissonDirichlet_Def.hpp. Now users only need to specify the applied voltage in the input xml. Currently only two types of contacts are supported, namely, contacts on insulators and ohmic contacts on semiconductors.
In the QCAD_PoissonSource_Def.hpp file, several small changes are made: remove scaling for the default device and set all states to 0 except electricPotential, remove the if...else statement in calculating qPhiRef since the same qPhiRef should be used for the entire structure, and minor change for the metal source term.

In the materials.xml file, several changes are made: change PolySilicon to Metal and change ElectronAffinity to Work Function for Metal, for the "substrate" nodeset use "elementBlock" instead of "material", add "dopantType" of "None" to the "silicon" element block. The rule for nodeset is specifying "material" for contacts on insulators and "elementBlock" for ohmic contacts on semiconductors.