Merging ASCR/UQ branch into master.
This merge adds several new capabilities:

* Basic implementation of stabilized (PSPG & SUPG) Navier-Stokes,
including two examples/tests.
* Some more SG solver approaches
* Some preliminary capabilities for adjoint-based a posteriori error
estimation.

If you get a build error in ENAT_SGNOXSolver.cpp after updating from this
commit, update Trilinos as a compiler-related bug was recently fixed.

This commit also fixed a bug in the Jacobian-scatter for multiphysics
problems (not all of the terms were being included if an offset was
specified for the scatter).  I had to rebaseline two tests after this
change.

One test is currently failing (Poisson_mosdot_2D_SERIAL), but that seems
unrelated to these changes.