Removed assumption that unknowns at a node are numbered sequentially. Throughout the code, there were lots of places with logic   ( id = node * neq + eq;), which have now been removed.
Note: None of this matters for Poisson or Schrodinger which
have 1 degree of freedom per node.

This choice  "Interleaved Ordering"=true is the default.
If this flag is set to false, then blocked ordering is done
instead.  For example, with 2 unknowns  u,v  at 3 nodes,
the interleaved ordering (default) is
  [u_0, v_0, u_1, v_1, u_2, v_2]
while the block ordering is
  [u_0, u_1, u_2, v_0, v_1, v_2].

Added 2 test problems that use block ordering,
NSVortexShedding2D_Blocked and NonlinearElasticity2D_Blocked