Add TimeManager class
This class has the purpose of keeping track of the simulation
(pseudo)time by grabbing the current_time from Rythmos or LOCA. Also
this allows the time step to be computed and could now be available for
the LAME models (TODO).

The benfit of this will be for implementating time dependent BCs. Other
benefits include the ability to write the current time to the (Exodus) output
file.

Next up, write some time dependent BCs as examples for general usage.