WARNING: This commit requires an update of Trilinos!
Added code in Albany and Trilinos so that Continuation
runs will send the parameter value for outputting in
the exodus time stamp. Eigenvectors have the associated
eigenvalue as the time stamp.

If this leads to non-(monitonically increasing) time
stamps, to see this in paraview you need to select
the "Has Mode Shapes" option.

Also in this commit are some changes to begin to
propagate Erik's response field manager approch
into the broader code.