Fixed pndiode problem to pass on both 1 and 4 procs.
I verified that I could get the same result if I
tightened the tolerances. The failure was on order
of 1.0e-10, but on a number of order 1.0e-8. Som
I added the Absolute Tolerance = 1.0e-6  line to the
Regression Results of the input file, which was
there for this situation.