Syncing Albany up with several recent changes in Stokhos and Piro.
This commit incorporates several changes/additions to Stokhos and Piro.
Primiarly ENAT_SGNOXSolver was removed and replaced by a new Piro Epetra
Stokhos solver.  At the same time as this commit, I also made several
other changes:

* Setup Albany to use an abritrary number of parameter vectors
(previously there was only one for deterministic problems and two for
stochastic problems).  Now you can specify how many you want in the
input file, although backwards compatability was maintained in the
case of one parameter vector.  The most significant impact of this for
stochastic problems is you no longer need to put in a deterministic
parameter vector to satisfy the ModelEvaluator logic.

* Added SGTangent and MPTangent evaluation types and the corresponding
compute functions in the Application.  This will enable stochastic
sensitivities.

* Filled in proper specializations for KfieldBC and
QCAD::EvaluationTools for the various AD types.

* Changed the signature of all of the Application compute functions in
terms of how they deal with parameters.

* Simplified the response function interface by removing the stochastic
Galerkin functions.  Instead the Application implements the SG
calculation for the responses through quadrature.  The impact of this is
any SG problem now has to use the Quadrature expansion method so that a
quadrature object is created.  This is only temporary until the response
field manager is finished and we will go back to truly embedded
evaluation of the responses.