Put all solver-related parameters in a "Piro" sublist.
I changed the structure of the input file to put all solver-related
parameter lists ("NOX", "LOCA", "Rythmos", etc...) in a "Piro" sublist.
This segregates these parameters from the rest of the Albany parameters
allowing Piro to validate its parameters better.  I also updated all of
the example/test input files for this change, and all of the tests pass.

*IMPORTANT*:  You must update any input xml files you use that aren't in
the examples directory.  Look at any one of the examples to see what to
do.