Forced time label sent to exdus to be monotonically increasing.
Paraview can not satisfactorily handle exodus data files when
the time value does not increase monotonically. When we use
parameter values or eigenvalues as the index, this causes
problems. So, I put in a small algorihm to select an increasing
time label even if the time value does not.

Here's the new screen output for these cases.

outputToExodus: writing time -16123 with label 16123 to index 2 in file wavefunctions3D.exo