Pushed modifications for multiple materials further through the code.
Changed output of 1D problems bac to using exodus. It turns out
that there is support for 1D problems in exodus (Thanks for letting
me know, Eric Cyr).  I just committed fixes to the  exotxt  applciation
in seacas so that this will dump 1D exodus files to a readable
format.