Set up plumbing in the build system to allow for unit testing.  I attempted to set things up in a way that is consistent with the existing CMakeLists.txt file structures.
Unit testing is set up to use the Teuchos unit test
harness.  Source code for individual tests is intended
to go in the new directory src/LCM/test/unit_tests.
The cmake directives to build the tests are located in
src/CMakeLists.txt, and the tests themselves will
be built in the src directory.  The cmake directives
to run the tests are located in the examples directory,
which is a bit of a misnomer in this case.  A UnitTests
directory was added under the examples directory.
A custom target "make utest" was created so that the
unit tests can be run separately from the full test
suite.

Currently there is one unit test, utLameStress_elastic.
This is only a placeholder at this point, it does not
actually test the LameStress evaluator.