Clean up NonlinearElasticity and Poro-Elasticity problems a bit (see below)
Strange behavior from the Albany::MDArray (which is just a shards::Array),
in that if I have a 4D array M(cell, qp, dim, dim), and I try to index into
it with the following types M( size_t, size_t, int, int), I get a compiler
error. Obvious workaround is not optimal, but it's not clear to me how to fix
this.