Progress toward unit testing the LameStress evaluator.
One difficulty in unit testing an evaluator is the need to
have all the required fields evaluated prior to calling
evaluateFields() on the evaluator you care about.  The
approach I've taken is to create, within the unit test,
an evaluator who's purpose is to fill in all the required
fields for the evaluator to be tested.  In the LameStress
unit test, this evaluator is currently called DefGradUnitTest;
it has no dependent fields and fills in the deformation gradient
per the unit test so that LameStress.evaluateFields() can be called
as directly as possible.  The plan is to generalize this and create
an evaluator called something like UnitTestRequiredFieldEvaluator
that fills in fields with explicitly supplied values from within
the unit test.

An alternative approach is to construct the full evaluation tree.
In the case of LameStress, this involves a half dozen evaluators,
which introduces unwanted and difficult-to-manage dependencies into
the unit test.