More progress toward unit testing the LameStress evaluator.
The plumbing is now set up so that LameStress::evaluateFields() can be
called.  Clean up work remains.

The way things are currently set up, the unit test must create an
Albany::StateManager prior to calling LameStress::evaluateFields().  Because
the StateManager is tied directly to STK data structures (I think), an object
deriving from Albany::AbstractDiscretization must be created and tied to
the StateManager.

The need to create an Albany::AbstractDiscretization has led to a large
number of dependencies for the unit test.  The test is currently creating a
discretization object via Albany::DiscretizationFactory.  Linking the
DiscretizationFactory code requires object files for seven different discretization
types, found in the stk directory.  In the spirit of unit testing, I'd like
to reduce the number of dependencies, so hopefully I can find a better way
of doing this, perhaps by either instantiating a discretization directoy (without
the discretization factory), or by creating a way to allocate state arrays in
the state manager without a discretization.