More progress on unit tests for the LameStress evaluator.
An initial unit test is in place that instantiates a LameStress evaluator
for the LAME elastic material model.  The evaluator is called with a
prescribed deformation gradient and the computed stress is compared against
the expected result.

Significant refactoring / clean-up is required to generalize the unit test
fixtures, and to add a series of non-trivial tests for a set of LAME
material models.