Add the beginnings of element block based material property support.
These changes begin to add support for element block based material
properties, in this case for thermal conductivity in the heat equation
problem(s). A parameter list validator was added to the PHAL_ThermalConductivity
evaluator which required minor changes to two examples to make them consistent.
The beginnings of a test case for cylindrical conduction through different materials
was also added. This is not hooked up yet.