Add ThermoMechanicalProblem as a knock off of the new ThermoElasticityProblem
ThermoMechanicalProblem currently has coupling between the temperature version
of the banalance of energy and the balance of linear momentum through finite
deformation kinematics and temperature dependent constitutive parameters.
Further extensions will include source terms in the energy equation due to
latent heat and internal dissipation.

Also, in an effort to simplify the kinematics interface, I completely the
LCG evaluator and finished off the removal of the RCG evaluator.

In the near term, I'd like to convert the other LCM problems over to the
factory-less design.