Add temperature dependence to bulk and shear moduli, isotropic thermal expansion
As a convenience, I added a reference temperature into the linear temperature
dependence of the bulk and shear modulus, so that the expression reads

k = k_0 + dkdT * ( Temp - refTemp )

since most of the time we use a linearized dependence about a certain temp. This
is alos used for thermal expansion.

Up next: refactor all (5) of the elastic constants into one evaluator.