Finished up example of multiple element blocks in 1D STK.
Finished example SteadyHeat1DEB that tests the generation of multiple element
block meshes in STK1D. Test was added to ctests. Note that this example also
employs the QCAD material database to look up thermal conductivities of different element
block materials. Lastly, there is a new regression test option "Solution Two Norm File"
that allows external generation of a reference solution (in this case generated by the
pre-test HeatProfile.C code), and the L2 comparison of this reference solution to that
calculated by the test.