Modified the GetEigensolverShift(...) function in QCAD_Solver.cpp to search the minimum energy only in quantum regions by using the condition of approxQuanEDen > 1e-5 because this quantity is only computed in quantum regions and is equal to 0 everywhere else. And the function is called starting from the 2nd S-P iteration as approxQuanEDen is not available during the 1st iteration.
Upload two Poisson-Schrodinger examples for Andy to debug two issues observed (already talked to Andy). One example is the PSParallelHang which is a 1D P-S example and hangs when trying to run it in parallel. Note: if using the GetEigensolverShift(...) function, each processor has a different shift which is not good as we want the same shift for the eigen solver; if disabling the GetEigensolverShift(...) function (comment out the two lines in QCAD_Solver.cpp), the same shift from input xml is used, but parallel run still hangs.

The second example is the PSPoissonConvIssue which is a 2D P-S example. During the S-P iterations, nonlinear Poisson solver would converge to a certain tolerance, say 1e-4, and then it remains at the value, but the nonlinear tolerance is specified as 1e-8 in the input xml, so Albany treats the Poisson solve as diverging and does not write responses to exo file. Andy suggested that it should be useful to investigate why the error of a Poisson solve stops at a certain value and not decreases further.