Created SetField evaluator and put it to use in the utLameStress_elastic unit test.
The SetField evaluator simply sets all the data in a specified
field to a given value.  This is useful in the testing of
evaluators which have a set of required fields.  The required
fields can be set directly with a SetField evaluator, allowing
the evaluator in question to be called without the need to
construct a full evaluation tree.