Fixed (hopefully) QCAD SaveField response function.
The save field response function was setup to only work with a residual
evaluation type, thus if the responses were called with a different
evaluation type, such as tangent (for dg/dp) for jacobian (for dg/dx),
then this response wouldn't do anything.  However if g and dg/dp or
dg/dx is requested, g won't be computed directly using the residual
evaluation type, rather along with dg/dp or dg/dx with the
tangent/jacobian evaluation type.  Thus this response wouldn't do its
thing in that case.  I fixed this by changing the response to work with
all evaluation types, however it only saves the value of the requested
field by extracting the value from the relevant scalar type.  Hopefully
this will fix Suzey's problem.