Fixed MPI communication of scalar types in responses.
** IMPORTANT:  You must updated Trilinos! **

I fixed the serialization objects to properly serialize empty AD
objects, which requires passing in the correct size of those objects to
the serializer.  I also moved the creation of the serializers out of the
gather solution revaluation and into the application.  This fixes
responses that don't depend on the solution.  Together these seem to get
the recent tests added by Erik and Suzey to work.