Added support for distributed response functions.
Distributed response functions are response functions with a non-local
map.  I added the respose function hierarchy for distributed responses,
and added two responses:  the (possibly culled) solution and a KL
expansion (the latter only works with SG currently).  To get this to
work I had to move the creation of the responses out of each individual
problem class and into the application.  I also modified a couple of the
examples to turn this on.  Finally I modified the drivers to be able to
deal with multiple responses and regression checks.  For the record, you
can specify multiple responses (any number of which can be distributed)
via:

    <ParameterList name="Response Functions">
      <Parameter name="Number of Response Vectors" type="int" value="2"/>
      <ParameterList name="Response Vector 0">
        <Parameter name="Name" type="string" value="Aggregated"/>
        <Parameter name="Number" type="int" value="1"/>
        <Parameter name="Response 0" type="string" value="PHAL Field Integral"/>
        <ParameterList name="ResponseParams 0">
          <Parameter name="Field Name" type="string" value="Temperature"/>
        </ParameterList>
      </ParameterList>
      <ParameterList name="Response Vector 1">
        <Parameter name="Name" type="string" value="Solution"/>
      </ParameterList>
    </ParameterList>

or, if there is only one scalar response per response vector, more simply:

    <ParameterList name="Response Functions">
      <Parameter name="Number of Response Vectors" type="int" value="2"/>
      <ParameterList name="Response Vector 0">
        <Parameter name="Name" type="string" value="PHAL Field Integral"/>
        <Parameter name="Field Name" type="string" value="Temperature"/>
      </ParameterList>
      <ParameterList name="Response Vector 1">
        <Parameter name="Name" type="string" value="Solution"/>
      </ParameterList>
    </ParameterList>