Bug fixes in QCAD Poisson-CI solver.  Interface between Albany and AlbanyCI tested on a small problem and results are qualitatively correct.  Only one CI call (no loop) has been tested, and particle number is still hardcoded to 2 for testing.
QCAD's Field Integral response now allows the user to specify the unit of the integrand (defaults to cm^-3 so no input files need to be changed).  This flexibility is used to fix unit errors in field integral responses used for the CI interface.

Note: Eigensolver shift has been set to be closer to the potential minimum than it was previously in order to get Anasazi to give correct eigenvectors in the example used for testing.  This may cause problems for Suzey's tests.  See comments in QCAD_Solver.cpp.