Enabled vector variables whose length is NOT the number of spatial dimensions.
Created separate "gradient" and "vector" data layouts, where we used
to just have "vector". gradient layouts have size numDims (spatial
dimension) while vector layouts have size vecDim (length of vector
of fields). Tensor data types are (vecDim,numDims) which is the
gradient of a vector field.

The Layout constructor, if no vecDim is given, will set these
two quantities to be equal and so there will be no difference
between the gradient and vector layouts. For the NavierStokes
AND ALL LCM PROBLEMS, I explicitly test that these are the
same. That is because these problems continue to use the
vector layouts for both vector fields and gradients of
scalar fields. So, these Problems are unchanged, but are
not ready to be run for cases when the number of
velocities or displacements is not equal to the spatial
dimension.

Adjusted the GPAM example problem (which solves N coupled species
balances) to have N=3 equations for a 2D mesh. This is the only
problem that uses or tests the new changes.