Added #defines and build options for linking with the Library of Advanced Materials for Engineering with Never-ending Templates (LAMENT).
Currently Albany can be built with either LAME or LAMENT, but not both.  The intension
is to enable building with both at some point in the future, such that the LAME stress
evaluator will make an AD call to a LAMENT material if available, and default to a
standard LAME material if a LAMENT material is not available.

The LAMENT library itself is currently a work in progress.