Added Elasticity2Traction test problem
Added traction to LCM/ElasticityProblem. Built up a test problem.
Specifics on the problem are contained in the TestNcdumpValues.C
code comments in the examples/Elasticity2DTraction directory. Worked
a bit on the transfer of field data to the exodus file - if one
has a steady state solution the field data for the final state of
registered fields did not get updated.