Progress toward a functioning LAMENT.
Added the function LameUtils::constructLameMaterialModel_AD() which returns a material
model with the template argument set to ADType.  ADType is currently typedefed to Sacado::Fad::DFad<double>
within LAMENT.  The getStress() function in a material model of this flavor takes an argument of type
matParams<ADType> in which strain_rate and stress_new are pointers to arrays of ADType and all other entries
are pointers to doubles (as in the standard LAME case).  This allows for calculation of the Jacobian for the
ElasticNew material model using the ADType (see the LAMENT unit test utElastic).

The next step is to use LameUtils::constructLameMaterialModel_AD() to construct a material model
in LameStress_Def and use that material model for calculation of the Jacobian via automatic differentiation.