Add ability to read a serial mesh and rebalance in parallel.
This commit adds capabilities to read a single Exodus mesh file on
PE 0 of a multiple PE run, and use STK rebalance (Zoltan) to distribute
the mesh across the machine. The test case is Ioss2D input_Serial.xml.
Work in progress was also committed: a) abilities for uniform refinement and
b) Albany_Adaptation class.