Added reduced-order model evaluator
For maximum flexibility and generality, the implementation is based on
the EpetraExt::ModelEvaluator interface, at the expense of performance.

The subclass ReducedOrderModelEvaluator wraps around any full-order model
that conforms the above interface.

Known limitations:
1) The Jacobian operator is not a RowMatrix, and thus is only usable
with solvers which do not require an assembled matrix. In particular,
it is not compatible with the Amesos package.
2) Only a limited subset of input/output arguments are supported
3) Only the Galerkin projection is implemented
4) The code has not been tested with distributed vectors