Reworked reduced-order Jacobian
The reduced-order Jacobian is now implemented as a CrsMatrix.
This change should enable the use of direct solvers of the Amesos package,
since they require operators that implement the RowMatrix interface.

Known limitations:
1) Most Amesos solvers do not support RowMatrices
whose domain and range maps are different from the row map,
and therefore cannot handle the generated reduced-order Jacobian.
2) Since the Jacobian is a dense matrix, the recommended direct solver
is Amesos_Lapack, but its use currently requires a bug fix
that is yet to be committed to the Trilinos development repository.
3) The reduced-order Jacobian implementation might violate some implicit invariants
of CrsMatrix that other Trilinos packages silently rely on.
4) Bugs whose appearance depend on the number of MPI processes
have been observed.